coding_test/number_5/pyproject.toml
[project]
name = "number-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "tesseract>=0.1.3",
    "litellm>=1.75.9",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "google-adk>=1.12.0",
    "streamlit>=1.48.1",
    "easyocr>=1.7.2",
]
